pyproject.toml
[project]
name = "httpfilegen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jsf>=0.11.2,<1.0",
    "openapi-pydantic>=0.5.1,<1.0",
    "openapi-spec-validator>=0.7.2,<1.0",
    "prance>=25.4.8.0,<26.0",
    "pydantic>=2.12.3,<3.0",
    "pydantic-settings>=2.11.0,<3.0",
    "typer>=0.20.0,<1.0",
]

[project.scripts]
httpfilegen = "cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["cli"]

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests", "tests.*", "*.tests", "*.tests.*"]

[project.optional-dependencies]
yaml = [
    "pyyaml>=6.0,<7.0",
]
test = [
    "pytest>=8.0,<9.0",
    "typer>=0.20.0,<1.0",
    "pytest-cov>=4.1,<6.0",
    "pyyaml>=6.0,<7.0",
]

[tool.pytest.ini_options]
addopts = "-q --cov=src --cov-report=term-missing --cov-report=xml:coverage.xml"
testpaths = ["tests"]
filterwarnings = [
    "ignore:Pickle, copy, and deepcopy support will be removed from itertools:DeprecationWarning:copy",
]

[tool.coverage.run]
branch = true
source = ["src"]
omit = [
    "*/tests/*",
    "*/.venv/*",
    "src/main.py",
]

[tool.coverage.report]
show_missing = true
skip_empty = true

[dependency-groups]
dev = [
    "ruff>=0.14.13",
]
